Add files via upload
simple batch script to launch 'sys_gen.py' python script that is supposed to generate the simulation system. This batch script assigns the HQ_TASK_ID parameter as an input argument to the 'sys_gen.py' python script which corresponds to the index number for the system in the list given in 'ligpairs.dat' file.