Add files via upload
An attempt to modify the 'Prepare_all.py' python script to be used as a library for 'sys_gen.py' python script. In the original 'Prepare_all.py' script, each system is defined as an edge which are collected in a list termed edges. Each 'build' function (for atom mapping, hybrid structure topology, assemble systems, box water ions) reads the edge from the list (in a serial fashion) and creates the system. This modified script (serial.py) will have the build functions modified such that it generates the system for only one 'edge' value. The particular 'edge' is assigned to the 'serial.py' script from the edges list based on the index value taken from HQ_TASK_ID.