Add files via upload
A modified version of the 'run.py' script. This script takes the input value from the 'batch.sh' script (HQ_TASK_ID) and assigns it as an index value for the 'edges' list. The corresponding 'edge' is then picked from the list to generate the system using the function provided in 'serial.py' script. Each submission of the 'run.py' script generates a single system from the 'ligpairs.dat' list.